path

+++ src/utils.rs
@@ -425,30 +425,30 @@ pub async fn get_owners_local(cache: Arc<Cache<String, f64>>) {
 
             let dir = dir.into_os_string().into_string().unwrap() + "/snapshots" + &filename;
             println!("{:?}", dir);
-            println!("{:?}", Path::new(&dir).exists());
-
-            let file = match std::fs::File::create(dir) {
-                Ok(x) => x,
-                Err(x) => {
-                    println!("createFileError: {:?}", x);
-                    continue;
-                }
-            };
-            let mut writer = BufWriter::new(file);
-            match serde_json::to_writer(&mut writer, &owners_map) {
-                Ok(x) => x,
-                Err(x) => {
-                    println!("SerdeToWritterError: {:?}", x);
-                    continue;
-                }
-            };
-            match writer.flush() {
-                Ok(x) => x,
-                Err(x) => {
-                    println!("FlushError: {:?}", x);
-                    continue;
-                }
-            };
+            if !Path::new(&dir).exists() {
+                let file = match std::fs::File::create(dir) {
+                    Ok(x) => x,
+                    Err(x) => {
+                        println!("createFileError: {:?}", x);
+                        continue;
+                    }
+                };
+                let mut writer = BufWriter::new(file);
+                match serde_json::to_writer(&mut writer, &owners_map) {
+                    Ok(x) => x,
+                    Err(x) => {
+                        println!("SerdeToWritterError: {:?}", x);
+                        continue;
+                    }
+                };
+                match writer.flush() {
+                    Ok(x) => x,
+                    Err(x) => {
+                        println!("FlushError: {:?}", x);
+                        continue;
+                    }
+                };
+            }
         }
         // thread::sleep(Duration::from_millis(300000));
     }
